% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajuste_poslabel.R
\name{ajuste_poslabel}
\alias{ajuste_poslabel}
\title{Ajusta las posiciones de las etiquetas para un gráfico de barras}
\usage{
ajuste_poslabel(data, posl, freq, xxn)
}
\arguments{
\item{data}{Tabla con los datos para un gráfico de barras}

\item{posl}{Vector con la posición de las etiquetas en las barras apiladas}

\item{freq}{Valores por graficar (porcentajes, usualmente)}

\item{xxn}{Opciones o categorias asociados a los porcentajes}
}
\value{
Agrega una columna 'poslabel_aj' a la tabla de datos
}
\description{
Con el objetivo de graficar barras apiladas, usualmente se calcula un vector con las posiciones de las etiquetas. Es posible que las etiquetas estes muy pegadas o se solapen, si es asi, la función agrega unos puntos a la posición de la etiqueta.
}
\examples{
opcion <- c("a", "b", "c", "d", "e")
freq = c(7, 46.7, 0.5, 1.4, 44.4)
pos_label = c(3.5, 30.35, 53.95, 54.9, 77.8)
bd <- data.frame(opcion = opcion, freq = freq, pos_label = pos_label)
bd1 <- ajuste_poslabel(bd, pos_label, freq, opcion)


}
