% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contar_nse_segregacion.R
\encoding{UTF-8}
\name{contar_nse_segregacion}
\alias{contar_nse_segregacion}
\title{Cuenta nse del estudiante según escuelas y clasifica segregación}
\usage{
contar_nse_segregacion(data, nse, codmod, area, c1, c2)
}
\arguments{
\item{data}{Base de datos}

\item{nse}{Variable que contiene el nivel socioeconómico, debe estar en números. Por ejemplo, 'nse alto' = 1, 'nse medio' = 2,...}

\item{codmod}{Variable que agrupa a los estudiantes, en principio la escuela (codmod7)}

\item{area}{Variable con el area de la escuela, debe contener 'Rural' y 'Urbano'}

\item{c1}{Corte para asignar segregación media}

\item{c2}{Corte para asignar segregación alta}
}
\value{

}
\description{
Se debe tener una base de datos con el nivel socioeconómico (NSE) de los estudiantes, el código modular y el area de la escuela. La función cuenta a los estudiantes en la escuela según su NSE (agrupando a los de NSE bajo y muy bajo) y despues asigna un nivel de segregación según los cortes c1 y c2. Finalmente, categoriza a las escuelas rurales.
}
\examples{
contar_nse_segregacion(data_ejemplo, nse, codmod7, area, 50, 90)

}
