% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_lista.R
\encoding{UTF-8}
\name{pega_lista}
\alias{pega_lista}
\title{Une dataframes que estan en una lista}
\usage{
pega_lista(data, nc)
}
\arguments{
\item{data}{Lista con dataframes}

\item{nc}{Asigna el nombre de la columna con los nombres de las listas}
}
\value{
Dataframe
}
\description{
En una lista de dataframes con la misma estructura, agrega una columna con el nombre de la columna y las une. Útil cuando cuando se hizo un 'split' por algún estrato.
}
\examples{
ejem <- list(hom = data.frame(v = 1:3), muj = data.frame(v = 4:6))
pega_lista(ejem, "nombre")


}
