% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_freq1.R
\name{tabla_freq1}
\alias{tabla_freq1}
\title{Igual que table(), pero en dataframe y porcentajes redondeados. Regresa un dataframe.}
\usage{
tabla_freq1(x)
}
\arguments{
\item{x}{Vector para aplicar tabla de frecuencias}
}
\value{
df
}
\description{
Igual que table(), pero en dataframe y porcentajes redondeados. Regresa un dataframe.
}
\examples{
tabla_freq1(iris$Species)
}
