% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_freq2.R
\name{tabla_freq2}
\alias{tabla_freq2}
\title{Tabla de frecuencias para varias columnas}
\usage{
tabla_freq2(data, x)
}
\arguments{
\item{data}{Base de datos}

\item{x}{Caracter con el que empieza el grupo de columnas}
}
\value{
DF
}
\description{
Tabla de frecuencias para varias columnas que empiezan con el mismo nombre y tienen el mismo contenido. Útil para un grupo de ítems de una escala, por ejemplo.
}
\examples{
tabla_freq2(data_ejemplo, "p02")

}
