% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_freq2_estrato.R
\name{tabla_freq2_estrato}
\alias{tabla_freq2_estrato}
\title{Tabla de frecuencias para varias columnas, según estrato}
\usage{
tabla_freq2_estrato(data, x, estrato)
}
\arguments{
\item{data}{Base de datos}

\item{x}{Caracter con el que empieza el grupo de columnas}

\item{estrato}{Nombre de la columna que contiene los estratos o grupos}
}
\value{

}
\description{
Tabla de frecuencias para varias columnas que empiezan con el mismo nombre y tienen el mismo contenido, agrupando por característica o estrato.
}
\examples{
tabla_freq2_estrato(data_ejemplo, "p02", "gestion")
}
